pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt_generator"
version = "0.1.0"
description = "一个基于AI的提示词生成和优化系统"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.8",
    "langchain-community>=0.0.28",
    "langchain-core>=0.1.31",
    "langgraph>=0.0.20",
    "openai>=1.14.2",
    "chromadb>=0.4.22",
    "faiss-cpu>=1.7.4",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.9.0",
    "black>=24.2.0",
    "isort>=5.13.2",
    "flake8>=7.0.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=html" 